Update package.json
This theme (BlueRexPY/SummerNight) is super cool! There's already an older theme also called Summer Night with several variants (github.com/jackw01/summer-night-vscode-theme or jackw01.summer-night-theme in the marketplace).

This theme is called Summer Night Pro in the marketplace but the label field in the package.json is "Summer Night", which is how it appears in the vs code themes list, creating a conflict if both are installed.

I'm suggesting a change to the label field in the package.json to clarify which theme is which when switching between themes. Summer Night (https://github.com/jackw01/summer-night-vscode-theme).